+++ src/V1/Checks/IsPluginProvider.php
@@ -1,0 +1,107 @@
+<?php
+
+/**
+ * Copyright (c) 2017-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   AdaptersAndPlugins\V1\Checks
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2017-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://ganbarodigital.github.io/php-mv-adapters-plugins
+ */
+
+namespace GanbaroDigital\AdaptersAndPlugins\V1\Checks;
+
+use GanbaroDigital\AdaptersAndPlugins\V1\Helpers;
+use GanbaroDigital\AdaptersAndPlugins\V1\PluginTypes\PluginClass;
+use GanbaroDigital\AdaptersAndPlugins\V1\PluginTypes\PluginProvider;
+use GanbaroDigital\MissingBits\Checks\Check;
+
+/**
+ * is an object a plugin provider?
+ */
+class IsPluginProvider implements Check
+{
+    /**
+     * is an object a plugin provider?
+     *
+     * @param  mixed $pluginProvider
+     *         the object to check
+     * @return boolean
+     *         - `true` if $pluginProvider is a plugin provider object
+     *         - `false` otherwise
+     */
+    public static function check($pluginProvider)
+    {
+        return ($pluginProvider instanceof PluginProvider);
+    }
+
+    /**
+     * is an object a plugin provider?
+     *
+     * @param  mixed $pluginProvider
+     *         the object to check
+     * @return boolean
+     *         - `true` if $pluginProvider is a plugin provider object
+     *         - `false` otherwise
+     */
+    public function __invoke($pluginProvider)
+    {
+        return self::check($pluginProvider);
+    }
+
+    /**
+     * is an object a plugin provider?
+     *
+     * @param  mixed $pluginProvider
+     *         the object to check
+     * @return boolean
+     *         - `true` if $pluginProvider is a plugin provider object
+     *         - `false` otherwise
+     */
+    public function inspect($pluginProvider)
+    {
+        return self::check($pluginProvider);
+    }
+
+    /**
+     * create a new, reusable Check object
+     *
+     * @return IsPluginProvider
+     */
+    public static function using()
+    {
+        return new self;
+    }
+}